Add support for Gutenberg
Create a new WordPress Gutenberg block to add a slider to the
page/post.

The block contains a simple dropdown menu listing the current sliders
in the database. When the user selects a slider in the menu, the
background of the block is set to the background of the first
slide of the selected slider.