Add compatibility with jQuery 3 and remove some deprecated JS code
Frontend and backend JS code generated various warnings or fatal
errors due to the deprecation of some functions in jQuery 3.
The fatal errors made the plugin incompatible with WordPress 5.6.

- Replaced jQuery mobile with Andrés Zsögön's jquery.events.swipe
library for handling swipe events on mobile and mouse drag events on
desktop. jQuery mobile is not compatible with jQuery 3, so it could no
longer be used.
- Replaced some jQuery functions that were still present in jQuery 3 but
considered deprecated.